Tests seperated to diffrenet files
