+++ backend/Tests/API_Test.py
@@ -1,6 +1,5 @@
 import socket
 import time
-import json
 
 IP = "127.0.0.1"
 PORT = 3391
@@ -27,12 +26,6 @@ def main():
     global totalRunTime
     
     print("Searching for server")
-    if discover_server():
-        print(f"{bcolors.OKGREEN} Server found {IP}:{str(PORT)} {bcolors.ENDC}") 
-
-    else:
-        print(f"{bcolors.FAIL} Failed to find server! {bcolors.ENDC}")
-        return -1
     
     print("=====================Starting API Tests=======================")
     TestAPI('{"Path":"/api/process/start", "Type":"UPDATE", "Data":"Minecraft Server"}')
@@ -73,47 +66,6 @@ def TestAPI(message):
         print(f"{bcolors.FAIL} Failed to message API: {err} {bcolors.ENDC}") 
 
     print("\n\n")
-
-def discover_server(timeout=5):
-    global IP
-    global PORT
-    global totalRunTime
-    
-    # Create a UDP socket
-    Succeeded = False
-    client_socket = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
-    client_socket.setsockopt(socket.SOL_SOCKET, socket.SO_BROADCAST, 1)
-    client_socket.settimeout(timeout)  # timeout in seconds
-
-    # Broadcast address
-    broadcast_address = ('192.168.50.255', DISCOVERY_PORT)
-
-    try:
-        # Send discovery message
-        client_socket.sendto(DISCOVERY_MESSAGE.encode('utf-8'), broadcast_address)
-        print(f"Broadcasted discovery request to {broadcast_address}")
-
-        # Wait for response from server
-        data, server_addr = client_socket.recvfrom(BUFFER_SIZE)
-
-        StartTime = time.perf_counter()
-        response = data.decode('utf-8')
-        elapsed_time = time.perf_counter() - StartTime
-        totalRunTime += elapsed_time
-
-        print(f"Received response: '{response}' from {server_addr[0]}:{server_addr[1]}")
-        print(f"{bcolors.OKCYAN}{bcolors.BOLD} Answer time: {(elapsed_time * 1000):.0f} ms{bcolors.ENDC}")
-
-        PORT = json.loads(response)["Port"]
-        IP = str(server_addr[0])
-        
-        Succeeded = True
-        
-    except socket.timeout:
-        print("No server response received within timeout.")
-    finally:
-        client_socket.close()
-    return Succeeded
 
 
 if __name__ == "__main__":
